updated notifications
