+++ itestify_backend/settings.py
@@ -82,8 +82,8 @@ GOOGLE_OAUTH_CLIENT_ID = os.getenv("GOOGLE_OAUTH_CLIENT_ID")
 GOOGLE_OAUTH_CLIENT_SECRET = os.getenv("GOOGLE_OAUTH_CLIENT_SECRET")
 GOOGLE_OAUTH_REDIRECT_URI = os.getenv("GOOGLE_OAUTH_REDIRECT_URI")
 
-#import warnings
-#warnings.filterwarnings("ignore", category=UserWarning, module='dj_rest_auth.registration.serializers')
+# import warnings
+# warnings.filterwarnings("ignore", category=UserWarning, module='dj_rest_auth.registration.serializers')
 
 
 # django-allauth (social)
